* Source/GSUnicode.c: Rework some of the string formatting code.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/corebase/trunk@38122 72102866-910b-0410-8b05-ffd578937521

+++ Source/CFString.c
@@ -523,7 +523,7 @@ CFShow (CFTypeRef obj)
 void
 CFShowStr (CFStringRef s)
 {
-  fprintf (stderr, "Length %ld\n", CFStringGetLength (s));
+  fprintf (stderr, "Length %ld\n", (long)CFStringGetLength (s));
   fprintf (stderr, "IsWide %d\n", CFStringIsUnicode (s));
   fprintf (stderr, "InlineContents %d\n", CFStringIsInline (s));
   fprintf (stderr, "Allocator %p\n", CFGetAllocator (s));
